email_mcp/requirements.txt
mcp>=1.2.0
google-api-python-client>=2.169.0
google-auth>=2.38.0
google-auth-oauthlib>=1.2.1
python-dotenv>=1.0.1
httpx>=0.28.1
